Migrate users components to a SingleForm layout
